update dependencies,move to moxy,inject DI

+++ app/src/main/java/com/arny/flightlogbook/data/models/AircraftType.kt
@@ -6,7 +6,7 @@ import android.arch.persistence.room.PrimaryKey
 import com.arny.flightlogbook.utils.Utility
 
 @Entity(tableName = "type_table")
-data class AircraftType(@PrimaryKey(autoGenerate = true) @ColumnInfo(name = "type_id") var typeId: Int = 0) {
+data class AircraftType(@PrimaryKey(autoGenerate = true) @ColumnInfo(name = "type_id") var typeId: Long = 0) {
     @ColumnInfo(name = "airplane_type")
     var typeName: String? = null
